fix: change BAUD rate field from int to long

+++ Tact/Tact.h
@@ -76,7 +76,7 @@
 			// Init Tact Toolkit
 			void begin();
 			// Init Tact Toolkit /w Serial
-			void beginSerial(unsigned int br = BAUD_RATE);
+			void beginSerial(long br = BAUD_RATE);
 			// Serial Event Delegation
 			void readSerial();
 			// Add single sensor with ID 0
